reimplemented the sample selection
