+++ mock_processing/samples/Flagship.py
@@ -1,0 +1,3 @@
+from . import reference
+
+from .base import SampleManager
