+++ mock_processing/samples/__init__.py
@@ -1,0 +1,6 @@
+from .base import SampleManager, DensitySampler, RedshiftSampler
+from .config import DumpConfig
+from . import reference  # must be imported
+# import all supported base simulations here
+from . import MICE2
+from . import Flagship
