+++ scripts/mocks_select_sample
@@ -6,19 +6,23 @@ import sys
 import mock_processing as mocks
 from mock_processing.core.bitmask import BitMaskManager
 from mock_processing.matching import DistributionEstimator
-from mock_processing import samples
+from mock_processing.samples import (DensitySampler, DumpConfig,
+                                     RedshiftSampler, SampleManager)
 
 
 parser = argparse.ArgumentParser(
     description="Apply sample selections on to obtain a mock representation "
                 "of a specific photometric or spectroscopic sample.",
     add_help=False)
-parser.register("action", "dump", samples.DumpConfig)
+parser.register("action", "dump", DumpConfig)
 
 parser.add_argument(
     "datastore", help="directory in which the data store is located")
 parser.add_argument(
-    "--sample", required=True, choices=sorted(samples.REGISTERED_SAMPLES),
+    "--type", default="reference", choices=SampleManager.flavours,
+    help="name of the sample selection to apply (default: %(default)s)")
+parser.add_argument(
+    "--sample", required=True, choices=SampleManager.samples,
     help="name of the sample selection to apply")
 parser.add_argument(
     "-c", "--config", required=True,
@@ -42,10 +46,10 @@ help_group.add_argument(
     "-h", "--help", action="help",
     help="show this help message and exit")
 help_group.add_argument(
-    "--dump", nargs=1, action="dump", metavar="sample",
-    choices=sorted(samples.SELECTION_PARSERS),
-    help="dump a default configuration file to stdout and exit for a given "
-         "sample (see --sample for a list of valid names)")
+    "--dump", action="dump",
+    help="dump a default configuration file for a combination of type and "
+         "sample to stdout and exit (see --type and --sample for a list of "
+         "valid names)")
 
 
 def main():
@@ -54,7 +58,7 @@ def main():
     logger = mocks.PipeLogger(__file__, args.datastore)
 
     # check the configuration file
-    Parser = samples.SELECTION_PARSERS[args.sample]
+    Parser = SampleManager.get_parser(args.type, args.sample)
     config = Parser(args.config, logger)
 
     # apply the magnification correction to the model magnitudes
@@ -77,7 +81,7 @@ def main():
 
         # start with photometric selection
         try:
-            selector_class = samples.PHOTOMETRIC_SELECTIONS[args.sample]
+            selector_class = SampleManager.get_selector(args.type, args.sample)
             selector = selector_class(BMM)
 
             ds.pool.set_worker(selector.apply)
@@ -95,39 +99,39 @@ def main():
             logger.handleException(e)
 
         # optional density sampling
-        try:
-            selector_class = samples.SAMPLING_SELECTIONS[args.sample]
+        sampler_class = SampleManager.get_sampler(args.type, args.sample)
+        if sampler_class is not NotImplemented:
+            try:
+                # surface density
+                if issubclass(sampler_class, DensitySampler):
+                    logger.info("estimating surface density ...")
+                    sampler = sampler_class(BMM, args.area, bitmask)
+                # redshift weighted density
+                else:
+                    logger.info("estimating redshift density ...")
+                    sampler = sampler_class(
+                        BMM, args.area, bitmask, ds[config.selection["redshift"]])
 
-            # surface density
-            if issubclass(selector_class, samples.DensitySampler):
-                logger.info("estimating surface density ...")
-                selector = selector_class(BMM, args.area, bitmask)
-            # redshift weighted density
-            else:
-                logger.info("estimating redshift density ...")
-                selector = selector_class(
-                    BMM, args.area, bitmask, ds[config.selection["redshift"]])
+                ds.pool.set_worker(sampler.apply)
+                # select the columns needed for the selection function
+                ds.pool.add_argument_column(config.bitmask)
+                # redshift weighted density requires a mock n(z) estimate
+                if isinstance(sampler, RedshiftSampler):
+                    ds.pool.add_argument_column(
+                        config.selection["redshift"], keyword="redshift")
 
-            ds.pool.set_worker(selector.apply)
-            # select the columns needed for the selection function
-            ds.pool.add_argument_column(config.bitmask)
-            # redshift weighted density requires a mock n(z) estimate
-            if isinstance(selector, samples.RedshiftSampler):
-                ds.pool.add_argument_column(
-                    config.selection["redshift"], keyword="redshift")
-
-            # apply selection
-            ds.pool.execute(seed=args.seed, prefix="density sampling")
-        except KeyError:
-            pass
-        except Exception as e:
-            logger.handleException(e)
+                # apply selection
+                ds.pool.execute(seed=args.seed, prefix="density sampling")
+            except KeyError:
+                pass
+            except Exception as e:
+                logger.handleException(e)
 
         # add the description attribute to the output column
         bitmask.attr = {"description": BMM.description}
 
         # show final statistics
-        N_mock = samples.DensitySampler.count_selected(bitmask)
+        N_mock = DensitySampler.count_selected(bitmask)
         density = "{:.3f} / arcmin²".format(N_mock / args.area / 3600.0)
         logger.info("density of selected objects: " + density)
         logger.info("computation completed for {:,d} entries".format(len(ds)))
